V2 solution
Added an updated solution that implements a state transition table as a data structure rather than the transitions explicitly coded.

Although the state table is much easier to modify, it runs on LeetCode.com twice as slow as the explicitly coded states.  This is most likely due to the fact that the state table is allocated for each tests.  Future iterations can see if LeetCode will allow me to statically allocate the state table once and reuse (i.e. singleton)